Deployment/requirements.txt
gunicorn
flask
tensorflow-cpu
numpy
matplotlib
seaborn
dill
scikit-image
pillow
opencv-python-headless
git+https://github.com/tensorflow/examples.git
segmentation-models
h5py
boto3